Add internal helpers package
