+++ crates/edr_napi/package.json
@@ -3,6 +3,7 @@
   "version": "0.5.2",
   "devDependencies": {
     "@napi-rs/cli": "^2.18.1",
+    "@nomicfoundation/edr-helpers": "workspace:*",
     "@types/chai": "^4.2.0",
     "@types/chai-as-promised": "^7.1.8",
     "@types/mocha": ">=9.1.0",
