+++ crates/edr_napi/test/solidity-tests.ts
@@ -1,37 +1,7 @@
 import { assert } from "chai";
+import { runAllSolidityTests } from "@nomicfoundation/edr-helpers";
 
-import {
-  ArtifactId,
-  ContractData,
-  SuiteResult,
-  runSolidityTests,
-  Artifact,
-  SolidityTestRunnerConfigArgs,
-} from "..";
-
-// This throws an error if the tests fail
-async function executeSolidityTests(
-  artifacts: Artifact[],
-  testSuites: ArtifactId[],
-  configArgs: SolidityTestRunnerConfigArgs
-): Promise<SuiteResult[]> {
-  return new Promise((resolve, reject) => {
-    const resultsFromCallback: SuiteResult[] = [];
-
-    runSolidityTests(
-      artifacts,
-      testSuites,
-      configArgs,
-      (result: SuiteResult) => {
-        resultsFromCallback.push(result);
-        if (resultsFromCallback.length === artifacts.length) {
-          resolve(resultsFromCallback);
-        }
-      },
-      reject
-    );
-  });
-}
+import { ArtifactId, ContractData, Artifact } from "..";
 
 describe("Solidity Tests", () => {
   it("executes basic tests", async function () {
@@ -45,7 +15,7 @@ describe("Solidity Tests", () => {
       projectRoot: __dirname,
     };
 
-    const results = await executeSolidityTests(artifacts, testSuites, config);
+    const results = await runAllSolidityTests(artifacts, testSuites, config);
 
     assert.equal(results.length, artifacts.length);
 
@@ -77,7 +47,7 @@ describe("Solidity Tests", () => {
     };
 
     await assert.isRejected(
-      executeSolidityTests(artifacts, testSuites, config),
+      runAllSolidityTests(artifacts, testSuites, config),
       Error
     );
   });
