+++ js/helpers/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@nomicfoundation/edr-helpers",
+  "description": "",
+  "version": "1.0.0",
+  "author": "",
+  "devDependencies": {
+    "@nomicfoundation/edr": "workspace:*",
+    "@tsconfig/node20": "^20.1.4",
+    "@types/node": "^20.0.0",
+    "@typescript-eslint/eslint-plugin": "5.61.0",
+    "@typescript-eslint/parser": "5.61.0",
+    "eslint": "^8.44.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-mocha": "10.4.1",
+    "eslint-plugin-prettier": "5.2.1",
+    "typescript": "~5.5.3"
+  },
+  "keywords": [],
+  "license": "MIT",
+  "main": "dist/index.js",
+  "private": true,
+  "scripts": {
+    "build": "tsc",
+    "eslint": "eslint 'src/**/*.ts'"
+  },
+  "types": "dist/index.d.ts"
+}
